src/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
python-socketio>=5.11.0
pydantic-settings>=2.1.0
psycopg2-binary>=2.9.9
scipy>=1.12.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
python-multipart>=0.0.9
requests>=2.31.0
pyserial>=3.5
redis>=5.0.1
celery>=5.3.6
PyYAML>=6.0.1
bcrypt>=4.1.2
python-jose[cryptography]>=3.3.0
supabase>=2.3.0
